+++ test/phoenix_live_view/integrations/hooks_test.exs
@@ -88,6 +88,48 @@ defmodule Phoenix.LiveView.HooksTest do
     assert lv |> element("#inc") |> render_click() =~ "count:8"
   end
 
+  test "handle_event/3 halts and replies", %{conn: conn} do
+    {:ok, lv, _html} = live(conn, "/lifecycle")
+
+    HooksLive.attach_hook(lv, :greet_1, :handle_event, fn "greet", %{"name" => name}, socket ->
+      {:halt, %{msg: "Hello, #{name}!"}, socket}
+    end)
+
+    HooksLive.attach_hook(lv, :greet_2, :handle_event, fn "greet", %{"name" => name}, socket ->
+      {:halt, %{msg: "Hi, #{name}!"}, socket}
+    end)
+
+    render_hook(lv, :greet, %{name: "Mike"})
+
+    assert_reply(lv, %{msg: "Hello, Mike!"})
+  end
+
+  test "only handle_event/3 error prints {:halt, map, %Socket{}}", %{conn: conn} do
+    {:ok, lv, _html} = live(conn, "/lifecycle")
+
+    HooksLive.attach_hook(lv, :boom, :handle_event, fn _, _, _ -> :boom end)
+
+    result =
+      HooksLive.exits_with(lv, ArgumentError, fn ->
+        lv |> element("#inc") |> render_click()
+      end)
+
+    assert result =~ "{:halt, map, %Socket{}}"
+    assert result =~ "Got: :boom"
+
+    {:ok, lv, _html} = live(conn, "/lifecycle")
+
+    HooksLive.attach_hook(lv, :reply, :handle_info, fn :boom, socket ->
+      {:halt, %{}, socket}
+    end)
+
+    assert ExUnit.CaptureLog.capture_log(fn ->
+             send(lv.pid, :boom)
+             ref = Process.monitor(lv.pid)
+             assert_receive {:DOWN, ^ref, _, _, _}
+           end) =~ "Got: {:halt, %{}, #Phoenix.LiveView.Socket<"
+  end
+
   test "handle_params/3 raises when hook result is invalid", %{conn: conn} do
     {:ok, lv, _html} = live(conn, "/lifecycle")